+++ src/main/java/org/opensearch/knn/index/engine/qframe/QuantizationConfig.java
@@ -8,6 +8,8 @@ package org.opensearch.knn.index.engine.qframe;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
+
+import org.opensearch.knn.index.engine.faiss.QFrameBitEncoder;
 import org.opensearch.knn.quantization.enums.ScalarQuantizationType;
 
 /**
@@ -19,5 +21,7 @@ import org.opensearch.knn.quantization.enums.ScalarQuantizationType;
 public class QuantizationConfig {
     @Builder.Default
     private ScalarQuantizationType quantizationType = null;
+    @Builder.Default
+    public boolean enableRandomRotation = QFrameBitEncoder.DEFAULT_ENABLE_RANDOM_ROTATION;
     public static final QuantizationConfig EMPTY = QuantizationConfig.builder().build();
 }