+++ src/test/java/org/opensearch/knn/index/quantizationservice/QuantizationServiceTests.java
@@ -20,10 +20,11 @@ import org.opensearch.knn.quantization.models.quantizationState.OneBitScalarQuan
 import org.opensearch.knn.quantization.models.quantizationState.QuantizationState;
 import java.io.IOException;
 import java.util.List;
+import java.util.function.Supplier;
 
 public class QuantizationServiceTests extends KNNTestCase {
     private QuantizationService<float[], byte[]> quantizationService;
-    private KNNVectorValues<float[]> knnVectorValues;
+    private Supplier<KNNVectorValues<float[]>> knnVectorValues;
 
     @Before
     public void setUp() throws Exception {
@@ -38,15 +39,20 @@ public class QuantizationServiceTests extends KNNTestCase {
         );
 
         // Use the predefined vectors to create KNNVectorValues
-        knnVectorValues = KNNVectorValuesFactory.getVectorValues(
+        // Use the predefined vectors to create KNNVectorValues
+        knnVectorValues = () -> KNNVectorValuesFactory.getVectorValues(
             VectorDataType.FLOAT,
             new TestVectorValues.PreDefinedFloatVectorValues(floatVectors)
         );
     }
 
     public void testTrain_oneBitQuantizer_success() throws IOException {
-        ScalarQuantizationParams oneBitParams = new ScalarQuantizationParams(ScalarQuantizationType.ONE_BIT);
-        QuantizationState quantizationState = quantizationService.train(oneBitParams, knnVectorValues, knnVectorValues.totalLiveDocs());
+        ScalarQuantizationParams oneBitParams = ScalarQuantizationParams.builder().sqType(ScalarQuantizationType.ONE_BIT).build();
+        QuantizationState quantizationState = quantizationService.train(
+            oneBitParams,
+            knnVectorValues,
+            knnVectorValues.get().totalLiveDocs()
+        );
 
         assertTrue(quantizationState instanceof OneBitScalarQuantizationState);
         OneBitScalarQuantizationState oneBitState = (OneBitScalarQuantizationState) quantizationState;
@@ -61,8 +67,12 @@ public class QuantizationServiceTests extends KNNTestCase {
     }
 
     public void testTrain_twoBitQuantizer_success() throws IOException {
-        ScalarQuantizationParams twoBitParams = new ScalarQuantizationParams(ScalarQuantizationType.TWO_BIT);
-        QuantizationState quantizationState = quantizationService.train(twoBitParams, knnVectorValues, knnVectorValues.totalLiveDocs());
+        ScalarQuantizationParams twoBitParams = ScalarQuantizationParams.builder().sqType(ScalarQuantizationType.TWO_BIT).build();
+        QuantizationState quantizationState = quantizationService.train(
+            twoBitParams,
+            knnVectorValues,
+            knnVectorValues.get().totalLiveDocs()
+        );
 
         assertTrue(quantizationState instanceof MultiBitScalarQuantizationState);
         MultiBitScalarQuantizationState multiBitState = (MultiBitScalarQuantizationState) quantizationState;
@@ -84,8 +94,12 @@ public class QuantizationServiceTests extends KNNTestCase {
     }
 
     public void testTrain_fourBitQuantizer_success() throws IOException {
-        ScalarQuantizationParams fourBitParams = new ScalarQuantizationParams(ScalarQuantizationType.FOUR_BIT);
-        QuantizationState quantizationState = quantizationService.train(fourBitParams, knnVectorValues, knnVectorValues.totalLiveDocs());
+        ScalarQuantizationParams fourBitParams = ScalarQuantizationParams.builder().sqType(ScalarQuantizationType.FOUR_BIT).build();
+        QuantizationState quantizationState = quantizationService.train(
+            fourBitParams,
+            knnVectorValues,
+            knnVectorValues.get().totalLiveDocs()
+        );
 
         assertTrue(quantizationState instanceof MultiBitScalarQuantizationState);
         MultiBitScalarQuantizationState multiBitState = (MultiBitScalarQuantizationState) quantizationState;
@@ -109,8 +123,12 @@ public class QuantizationServiceTests extends KNNTestCase {
     }
 
     public void testQuantize_oneBitQuantizer_success() throws IOException {
-        ScalarQuantizationParams oneBitParams = new ScalarQuantizationParams(ScalarQuantizationType.ONE_BIT);
-        QuantizationState quantizationState = quantizationService.train(oneBitParams, knnVectorValues, knnVectorValues.totalLiveDocs());
+        ScalarQuantizationParams oneBitParams = ScalarQuantizationParams.builder().sqType(ScalarQuantizationType.ONE_BIT).build();
+        QuantizationState quantizationState = quantizationService.train(
+            oneBitParams,
+            knnVectorValues,
+            knnVectorValues.get().totalLiveDocs()
+        );
 
         QuantizationOutput quantizationOutput = quantizationService.createQuantizationOutput(oneBitParams);
 
@@ -124,8 +142,12 @@ public class QuantizationServiceTests extends KNNTestCase {
     }
 
     public void testQuantize_twoBitQuantizer_success() throws IOException {
-        ScalarQuantizationParams twoBitParams = new ScalarQuantizationParams(ScalarQuantizationType.TWO_BIT);
-        QuantizationState quantizationState = quantizationService.train(twoBitParams, knnVectorValues, knnVectorValues.totalLiveDocs());
+        ScalarQuantizationParams twoBitParams = ScalarQuantizationParams.builder().sqType(ScalarQuantizationType.TWO_BIT).build();
+        QuantizationState quantizationState = quantizationService.train(
+            twoBitParams,
+            knnVectorValues,
+            knnVectorValues.get().totalLiveDocs()
+        );
         QuantizationOutput quantizationOutput = quantizationService.createQuantizationOutput(twoBitParams);
         byte[] quantizedVector = quantizationService.quantize(quantizationState, new float[] { 4.0f, 5.0f, 6.0f }, quantizationOutput);
 
@@ -137,8 +159,12 @@ public class QuantizationServiceTests extends KNNTestCase {
     }
 
     public void testQuantize_fourBitQuantizer_success() throws IOException {
-        ScalarQuantizationParams fourBitParams = new ScalarQuantizationParams(ScalarQuantizationType.FOUR_BIT);
-        QuantizationState quantizationState = quantizationService.train(fourBitParams, knnVectorValues, knnVectorValues.totalLiveDocs());
+        ScalarQuantizationParams fourBitParams = ScalarQuantizationParams.builder().sqType(ScalarQuantizationType.FOUR_BIT).build();
+        QuantizationState quantizationState = quantizationService.train(
+            fourBitParams,
+            knnVectorValues,
+            knnVectorValues.get().totalLiveDocs()
+        );
         QuantizationOutput quantizationOutput = quantizationService.createQuantizationOutput(fourBitParams);
 
         byte[] quantizedVector = quantizationService.quantize(quantizationState, new float[] { 7.0f, 8.0f, 9.0f }, quantizationOutput);
@@ -151,8 +177,12 @@ public class QuantizationServiceTests extends KNNTestCase {
     }
 
     public void testQuantize_whenInvalidInput_thenThrows() throws IOException {
-        ScalarQuantizationParams oneBitParams = new ScalarQuantizationParams(ScalarQuantizationType.ONE_BIT);
-        QuantizationState quantizationState = quantizationService.train(oneBitParams, knnVectorValues, knnVectorValues.totalLiveDocs());
+        ScalarQuantizationParams oneBitParams = ScalarQuantizationParams.builder().sqType(ScalarQuantizationType.ONE_BIT).build();
+        QuantizationState quantizationState = quantizationService.train(
+            oneBitParams,
+            knnVectorValues,
+            knnVectorValues.get().totalLiveDocs()
+        );
         QuantizationOutput quantizationOutput = quantizationService.createQuantizationOutput(oneBitParams);
         assertThrows(IllegalArgumentException.class, () -> quantizationService.quantize(quantizationState, null, quantizationOutput));
     }