+++ src/test/java/org/opensearch/knn/quantization/quantizationState/QuantizationStateSerializerTests.java
@@ -16,25 +16,35 @@ import java.io.IOException;
 public class QuantizationStateSerializerTests extends KNNTestCase {
 
     public void testSerializeAndDeserializeOneBitScalarQuantizationState() throws IOException {
-        ScalarQuantizationParams params = new ScalarQuantizationParams(ScalarQuantizationType.ONE_BIT);
+        ScalarQuantizationParams params = ScalarQuantizationParams.builder().sqType(ScalarQuantizationType.ONE_BIT).build();
         float[] mean = new float[] { 0.1f, 0.2f, 0.3f };
-        OneBitScalarQuantizationState state = new OneBitScalarQuantizationState(params, mean);
+        float[][] rotationMatrix = new float[][] { { 1.0f, 0.0f, 0.0f }, { 0.0f, 1.0f, 0.0f }, { 0.0f, 0.0f, 1.0f } };
 
-        // Serialize
+        OneBitScalarQuantizationState state = OneBitScalarQuantizationState.builder()
+            .quantizationParams(params)
+            .meanThresholds(mean)
+            .rotationMatrix(rotationMatrix)
+            .build();
+
         byte[] serialized = state.toByteArray();
-
         OneBitScalarQuantizationState deserialized = OneBitScalarQuantizationState.fromByteArray(serialized);
 
         assertArrayEquals(mean, deserialized.getMeanThresholds(), 0.0f);
         assertEquals(params, deserialized.getQuantizationParams());
+        assertArrayEquals(rotationMatrix[0], deserialized.getRotationMatrix()[0], 0.0f);
     }
 
     public void testSerializeAndDeserializeMultiBitScalarQuantizationState() throws IOException {
-        ScalarQuantizationParams params = new ScalarQuantizationParams(ScalarQuantizationType.TWO_BIT);
+        ScalarQuantizationParams params = ScalarQuantizationParams.builder().sqType(ScalarQuantizationType.TWO_BIT).build();
         float[][] thresholds = new float[][] { { 0.1f, 0.2f, 0.3f }, { 0.4f, 0.5f, 0.6f } };
-        MultiBitScalarQuantizationState state = new MultiBitScalarQuantizationState(params, thresholds);
+        float[][] rotationMatrix = new float[][] { { 1.0f, 0.0f, 0.0f }, { 0.0f, 1.0f, 0.0f }, { 0.0f, 0.0f, 1.0f } };
 
-        // Serialize
+        MultiBitScalarQuantizationState state = MultiBitScalarQuantizationState.builder()
+            .quantizationParams(params)
+            .thresholds(thresholds)
+            .rotationMatrix(rotationMatrix)
+            .build();
+
         byte[] serialized = state.toByteArray();
         MultiBitScalarQuantizationState deserialized = MultiBitScalarQuantizationState.fromByteArray(serialized);
 
@@ -42,5 +52,6 @@ public class QuantizationStateSerializerTests extends KNNTestCase {
             assertArrayEquals(thresholds[i], deserialized.getThresholds()[i], 0.0f);
         }
         assertEquals(params, deserialized.getQuantizationParams());
+        assertArrayEquals(rotationMatrix[1], deserialized.getRotationMatrix()[1], 0.0f);
     }
 }